+++ core/src/main/java/io/grpc/InternalChannelz.java
@@ -22,9 +22,6 @@ import static com.google.common.base.Preconditions.checkState;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.MoreObjects;
 import com.google.common.base.Objects;
-import io.grpc.internal.Instrumented;
-import io.grpc.internal.LogId;
-import io.grpc.internal.WithLogId;
 import java.net.SocketAddress;
 import java.security.cert.Certificate;
 import java.util.ArrayList;
