+++ core/src/main/java/io/grpc/internal/AbstractSubchannel.java
@@ -17,6 +17,7 @@
 package io.grpc.internal;
 
 import com.google.common.annotations.VisibleForTesting;
+import io.grpc.Instrumented;
 import io.grpc.InternalChannelz.ChannelStats;
 import io.grpc.LoadBalancer;
 import javax.annotation.Nullable;
