+++ core/src/main/java/io/grpc/internal/DelayedClientTransport.java
@@ -25,6 +25,7 @@ import io.grpc.InternalChannelz.SocketStats;
 import io.grpc.LoadBalancer.PickResult;
 import io.grpc.LoadBalancer.PickSubchannelArgs;
 import io.grpc.LoadBalancer.SubchannelPicker;
+import io.grpc.LogId;
 import io.grpc.Metadata;
 import io.grpc.MethodDescriptor;
 import io.grpc.Status;
