+++ core/src/main/java/io/grpc/internal/ServerImpl.java
@@ -36,9 +36,11 @@ import io.grpc.Context;
 import io.grpc.Decompressor;
 import io.grpc.DecompressorRegistry;
 import io.grpc.HandlerRegistry;
+import io.grpc.Instrumented;
 import io.grpc.InternalChannelz;
 import io.grpc.InternalChannelz.ServerStats;
 import io.grpc.InternalServerInterceptors;
+import io.grpc.LogId;
 import io.grpc.Metadata;
 import io.grpc.ServerCall;
 import io.grpc.ServerCallHandler;
