+++ core/src/test/java/io/grpc/internal/ServerImplTest.java
@@ -54,11 +54,13 @@ import io.grpc.Compressor;
 import io.grpc.Context;
 import io.grpc.Grpc;
 import io.grpc.HandlerRegistry;
+import io.grpc.Instrumented;
 import io.grpc.IntegerMarshaller;
 import io.grpc.InternalChannelz;
 import io.grpc.InternalChannelz.ServerSocketsList;
 import io.grpc.InternalChannelz.SocketStats;
 import io.grpc.InternalServerInterceptors;
+import io.grpc.LogId;
 import io.grpc.Metadata;
 import io.grpc.MethodDescriptor;
 import io.grpc.ServerCall;
