+++ grpclb/src/main/java/io/grpc/grpclb/GrpclbLoadBalancer.java
@@ -23,13 +23,13 @@ import io.grpc.Attributes;
 import io.grpc.ConnectivityStateInfo;
 import io.grpc.EquivalentAddressGroup;
 import io.grpc.LoadBalancer;
+import io.grpc.LogId;
 import io.grpc.Status;
+import io.grpc.WithLogId;
 import io.grpc.internal.BackoffPolicy;
 import io.grpc.internal.GrpcAttributes;
-import io.grpc.internal.LogId;
 import io.grpc.internal.ObjectPool;
 import io.grpc.internal.TimeProvider;
-import io.grpc.internal.WithLogId;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
