+++ netty/src/main/java/io/grpc/netty/NettyServerTransport.java
@@ -23,9 +23,9 @@ import com.google.common.collect.ImmutableList;
 import com.google.common.util.concurrent.ListenableFuture;
 import com.google.common.util.concurrent.SettableFuture;
 import io.grpc.InternalChannelz.SocketStats;
+import io.grpc.LogId;
 import io.grpc.ServerStreamTracer;
 import io.grpc.Status;
-import io.grpc.internal.LogId;
 import io.grpc.internal.ServerTransport;
 import io.grpc.internal.ServerTransportListener;
 import io.grpc.internal.TransportTracer;
