+++ okhttp/src/test/java/io/grpc/okhttp/OkHttpClientTransportTest.java
@@ -55,6 +55,7 @@ import com.google.common.util.concurrent.Futures;
 import com.google.common.util.concurrent.MoreExecutors;
 import com.google.common.util.concurrent.SettableFuture;
 import io.grpc.CallOptions;
+import io.grpc.Instrumented;
 import io.grpc.InternalChannelz.SocketStats;
 import io.grpc.InternalChannelz.TransportStats;
 import io.grpc.InternalStatus;
@@ -68,7 +69,6 @@ import io.grpc.internal.AbstractStream;
 import io.grpc.internal.ClientStreamListener;
 import io.grpc.internal.ClientTransport;
 import io.grpc.internal.GrpcUtil;
-import io.grpc.internal.Instrumented;
 import io.grpc.internal.ManagedClientTransport;
 import io.grpc.internal.ProxyParameters;
 import io.grpc.internal.TransportTracer;
