+++ services/src/test/java/io/grpc/services/ChannelzTestHelper.java
@@ -20,6 +20,7 @@ import com.google.common.base.MoreObjects;
 import com.google.common.util.concurrent.ListenableFuture;
 import com.google.common.util.concurrent.SettableFuture;
 import io.grpc.ConnectivityState;
+import io.grpc.Instrumented;
 import io.grpc.InternalChannelz;
 import io.grpc.InternalChannelz.ChannelStats;
 import io.grpc.InternalChannelz.Security;
@@ -27,9 +28,8 @@ import io.grpc.InternalChannelz.ServerStats;
 import io.grpc.InternalChannelz.SocketOptions;
 import io.grpc.InternalChannelz.SocketStats;
 import io.grpc.InternalChannelz.TransportStats;
-import io.grpc.internal.Instrumented;
-import io.grpc.internal.LogId;
-import io.grpc.internal.WithLogId;
+import io.grpc.LogId;
+import io.grpc.WithLogId;
 import java.net.InetSocketAddress;
 import java.net.SocketAddress;
 import java.util.Collections;
